+++ src/GUI/GUI.java
@@ -1,4 +1,4 @@
-package GUI;
+package src.GUI;
 
 public abstract class GUI {
