+++ src/GUI/MainWindow.java
@@ -1,0 +1,17 @@
+package src.GUI;
+
+import javax.swing.*;
+
+public class MainWindow {
+    private JButton co_btn;
+    private JList list1;
+    private JPanel co_usr;
+    private JTextField msg_input;
+    private JButton send_btn;
+    private JButton attachFile_btn;
+    private JTextArea textArea1;
+
+    private void createUIComponents() {
+        // TODO: place custom component creation code here
+    }
+}
